fix(rust): resolve clippy errors and warnings in pcai_core_lib
Fixes:
- Add tempfile as dev-dependency for test modules
- Allow clippy::not_unsafe_ptr_arg_deref for FFI functions (documented pattern)
- Fix boolean logic bug in is_windows_path()
- Remove unused imports in hash.rs
- Fix redundant field name in usb.rs
- Add #[default] derive for PcaiStatus enum
- Add #[allow(dead_code)] for reserved Hyper-V structs
- Remove unnecessary i32 cast for SOCK_STREAM
- Apply clippy auto-fixes for map_or, div_ceil patterns

Build now passes with 8 remaining documentation warnings.

+++ Native/pcai_core/pcai_core_lib/src/vmm_health.rs
@@ -1,9 +1,16 @@
+//! VMM Health Module - Hyper-V socket connectivity checks
+//!
+//! Provides health checks for WSL2/Hyper-V host communication.
+//! Some structures and constants are reserved for future direct vsock support.
+
 use windows_sys::Win32::Networking::WinSock::{
     socket, AF_HYPERV, SOCK_STREAM, INVALID_SOCKET,
 };
 use windows_sys::core::GUID;
 
 // Define SOCKADDR_HV manually since it might not be in windows-sys's version for networking
+// Reserved for future direct vsock connect support
+#[allow(dead_code)]
 #[repr(C)]
 #[derive(Clone, Copy)]
 struct SockAddrHv {
@@ -13,7 +20,8 @@ struct SockAddrHv {
     service_id: GUID,
 }
 
-// Well-known GUIDs
+// Well-known Hyper-V GUIDs - reserved for future vsock connect support
+#[allow(dead_code)]
 const HV_GUID_PARENT: GUID = GUID {
     data1: 0xa42e7cd,
     data2: 0x5542,
@@ -21,6 +29,7 @@ const HV_GUID_PARENT: GUID = GUID {
     data4: [0xaa, 0x44, 0x89, 0x6e, 0x99, 0xa9, 0x7a, 0x2b],
 };
 
+#[allow(dead_code)]
 const HV_GUID_CHILDREN: GUID = GUID {
     data1: 0x9092221,
     data2: 0x2728,
@@ -28,6 +37,7 @@ const HV_GUID_CHILDREN: GUID = GUID {
     data4: [0x91, 0x84, 0x20, 0xfa, 0x1d, 0x20, 0xa2, 0x06],
 };
 
+#[allow(dead_code)]
 const HV_GUID_LOOPBACK: GUID = GUID {
     data1: 0xe0e4ca1,
     data2: 0x496a,
@@ -35,6 +45,7 @@ const HV_GUID_LOOPBACK: GUID = GUID {
     data4: [0xac, 0x27, 0x4a, 0x71, 0x90, 0xc4, 0x1b, 0x27],
 };
 
+#[allow(dead_code)]
 const HV_GUID_WILDCARD: GUID = GUID {
     data1: 0x00000000,
     data2: 0x0000,
@@ -58,7 +69,7 @@ pub fn check_vmm_health() -> VmmHealth {
     };
 
     unsafe {
-        let s = socket(AF_HYPERV as i32, SOCK_STREAM as i32, 0);
+        let s = socket(AF_HYPERV as i32, SOCK_STREAM, 0);
         if s != INVALID_SOCKET {
             health.vsock_available = true;
             // For now, we just check if we can create the socket.
